requirements.txt
flask==2.3.3
tensorflow==2.13.0
pillow==10.0.0
numpy==1.24.3
flask-cors==4.0.0
opencv-python==4.8.0.76
matplotlib==3.7.2
requests==2.31.0
werkzeug==2.3.7